requirements.txt
fastapi>=0.100.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
